SDK build script does not crash if appledoc is not installed

+++ ShortcutSDK/BuildScripts/render_docs.sh
@@ -6,19 +6,26 @@ companyURL="http://shortcutmedia.com";
 outputPath="${BUILT_PRODUCTS_DIR}/docs";
 
 rm -rf "${outputPath}"
+mkdir -p "${outputPath}"
 
-/usr/local/bin/appledoc \
---project-name "${PROJECT_NAME}" \
---project-company "${company}" \
---company-id "#{companyID}" \
---output "${outputPath}" \
---logformat xcode \
---keep-intermediate-files \
---no-repeat-first-par \
---no-warn-invalid-crossref \
---exit-threshold 2 \
---create-html \
-"${PROJECT_DIR}"
+if [ -x /usr/local/bin/appledoc ]; then
 
-rm -rf "${outputPath}/docset"
-rm -f "${outputPath}/docset-installed.txt"
+    /usr/local/bin/appledoc \
+      --project-name "${PROJECT_NAME}" \
+      --project-company "${company}" \
+      --company-id "#{companyID}" \
+      --output "${outputPath}" \
+      --logformat xcode \
+      --keep-intermediate-files \
+      --no-repeat-first-par \
+      --no-warn-invalid-crossref \
+      --exit-threshold 2 \
+      --create-html \
+      "${PROJECT_DIR}"
+
+    rm -rf "${outputPath}/docset"
+    rm -f "${outputPath}/docset-installed.txt"
+
+fi
+
+open "${outputPath}"
